Module Otfm.Tag

Tags.

OpenType tags are four bytes identifiers.

Tags

type t = tag

The type for OpenType tags.

Table tags

Required tables

val cmap : tag

The cmap table.

val head : tag

The cmap table.

The head table.

val hhea : tag

The head table.

The hhea table.

val hmtx : tag

The hhea table.

The hmtx table.

val maxp : tag

The hmtx table.

The maxp table.

val name : tag

The maxp table.

The name table.

val os2 : tag

The name table.

The os2 table.

val post : tag

The os2 table.

The post table.

TTF font tables

val cvt : tag

The cvt table.

val fpgm : tag

The cvt table.

The fpgm table.

val glyf : tag

The fpgm table.

The glyf table.

val loca : tag

The glyf table.

The loca table.

val prep : tag

The loca table.

The prep table.

CFF font tables

val cff : tag

The CFF table.

val vorg : tag

The CFF table.

The VORG table.

Bitmap glyph tables

val ebdt : tag

The EBDT table.

val eblc : tag

The EBLC table.

val ebsc : tag

The EBSC table.

Optional tables

val dsig : tag

The DSIG table.

val gasp : tag

The DSIG table.

The gasp table.

val hdmx : tag

The gasp table.

The hdmx table.

val kern : tag

The hdmx table.

The kern table.

val ltsh : tag

The kern table.

The LTSH table.

val pclt : tag

The LTSH table.

The PCLT table.

val vdmx : tag

The PCLT table.

The VDMX table.

val vhea : tag

The VDMX table.

The vhea table.

val vmtx : tag

The vhea table.

The vmtx table.

Advanced typographic tables

val base : tag

The BASE table.

val gdef : tag

The BASE table.

The GDEF table.

val gpos : tag

The GDEF table.

The GPOS table.

val gsub : tag

The GPOS table.

The GSUB table.

val jstf : tag

The GSUB table.

The JSTF table.

Functions

val of_bytes : string -> tag

of_bytes s is a tag corresponding to s.

raises Invalid_argument
if s is not four byte long.

val to_bytes : tag -> string

to_string t is the tag as a four byte long string.

val to_int32 : tag -> int32

to_int32 t is the tag as an unsigned 32 bits integer.

val of_int32 : int32 -> tag

of_int32 t is the tag from and unsigned 32 bits integer.

val compare : tag -> tag -> int

compare t t' is Pervasives.compare t t'

val pp : Format.formatter -> tag -> unit

pp t prints a textual representation of t on ppf.